+++ src/main/java/com/epic60869/skyjew/SkyJewMod.java
@@ -66,6 +66,7 @@ public final class SkyJewMod implements ClientModInitializer {
         com.epic60869.skyjew.features.misc.PricePaid.init(configDir);
         com.epic60869.skyjew.features.misc.CollectionTracker.init(configDir);
         SkyJewStaff.init();
+        SkyJewUpdateChecker.init();
         SkyJewChangelog.init();
         SkyJewNopoFeatures.init(configDir);
         SkyJewNick.init(config);